requirements.txt
fastapi
uvicorn
google-cloud-storage
google-cloud-aiplatform
circleci
docker
pydantic